Addressed the circular imports
